monorepo/libs/tdp-fastapi-core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tdp-fastapi-core"
version = "0.1.0"
description = "Shared platform library for TDP FastAPI services"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.100.0,<1",
    "python-json-logger>=2.0.0,<3",
]

[project.optional-dependencies]
telemetry = [
    "opentelemetry-api>=1.20.0,<2",
    "opentelemetry-sdk>=1.20.0,<2",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0,<2",
    "opentelemetry-instrumentation-fastapi>=0.41b0,<1",
]
telemetry-httpx = [
    "opentelemetry-instrumentation-httpx>=0.41b0,<1",
]
telemetry-sqlalchemy = [
    "opentelemetry-instrumentation-sqlalchemy>=0.41b0,<1",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "httpx>=0.24",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "opentelemetry-instrumentation-httpx>=0.41b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.41b0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/tdp_fastapi_core"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
